requirement.txt
fastapi
uvicorn[standard]
pydantic
google-genai
python-dotenv
requests